+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.4.0] / 2024-09-23
 ### Shapes
 - Add `BRepBuilderExtension` with extension for face and solid.
+### Tests
+- Add `Shape_BRepBuilder_Tests` to test `BRepBuilderExtension`.
 
 ## [0.3.2] / 2024-06-14
 ### Shapes